Fix error handling for failed image conversions
The image inspector introduced with the fix for
OSSA-2024-001/CVE-2024-32498 just throws RuntimeException on error. But
this execption was not recognized as a image task failure and thus the
image status was not changed back to "queued".

This is fixed by reseting the status in all cases the task is reverted.

The same change was made upstream as part of
https://review.opendev.org/c/openstack/glance/+/840318 which implements
the "glance-download" import method.

This fix can be dropped with Zed which contains the upstream fix.